+++ platform/gl/gl-file.c
@@ -211,7 +211,7 @@ static void load_dir(const char *path)
 	}
 	else
 	{
-		while ((dp = readdir(dir)) && fc.count < nelem(fc.files))
+		while ((dp = readdir(dir)) && fc.count < (int)nelem(fc.files))
 		{
 			/* skip hidden files */
 			if (dp->d_name[0] == '.' && strcmp(dp->d_name, ".") && strcmp(dp->d_name, ".."))
@@ -282,7 +282,7 @@ static void list_drives(void)
 
 	ui_list_begin(&drive_list, nelem(common_dirs), 0, nelem(common_dirs) * ui.lineheight + 4);
 
-	for (i = 0; i < nelem(common_dirs); ++i)
+	for (i = 0; i < (int)nelem(common_dirs); ++i)
 		if (has_dir(home, user, i, dir, vis))
 			if (ui_list_item(&drive_list, common_dirs[i].name, vis, 0))
 				load_dir(dir);
