+++ source/fitz/load-gif.c
@@ -423,7 +423,7 @@ gif_read_ae(fz_context *ctx, struct info *info, const unsigned char *p, const un
 		fz_throw(ctx, FZ_ERROR_GENERIC, "out of range application extension block size in gif image");
 
 	ignored = 0;
-	for (i = 0; i < nelem(ignorable); i++)
+	for (i = 0; i < (int)nelem(ignorable); i++)
 		ignored |= memcmp(&p[3], ignorable[i], 8 + 3);
 	if (!ignored)
 	{
