+++ source/xps/xps-common.c
@@ -102,7 +102,7 @@ xps_begin_opacity(fz_context *ctx, xps_document *doc, fz_matrix ctm, fz_rect are
 		opacity_mask_tag = NULL;
 	}
 
-	if (doc->opacity_top + 1 < nelem(doc->opacity))
+	if (doc->opacity_top + 1 < (int)nelem(doc->opacity))
 	{
 		doc->opacity[doc->opacity_top + 1] = doc->opacity[doc->opacity_top] * opacity;
 		doc->opacity_top++;
